+++ res/views/components/label.blade.php
@@ -1,7 +1,3 @@
-<label @isset($for)for="{{ $for }}"@endisset>
-    @if ($text)
-    <x-ui::text :value="$text"/>
-    @else
-    {{ $slot }}
-    @endif
-</label>
+@ui('label', $attributes)
+    <s-text :value="$text ?? $slot"/>
+@endui