+++ res/views/components/link.blade.php
@@ -2,10 +2,6 @@
 $href = isset($route) ? route($route, $params) : url($href);
 @endphp
 
-@ui('a', $attributes->merge(['href' => $href]))
-    @isset($text)
-    <x-ui::text :value="$text"/>
-    @else
-    {{ $slot }}
-    @endisset
+@ui('a', $attributes->merge(compact('href')))
+    @s('text', ['value' => $text], $slot)
 @endui