+++ src/Components/Text.php
@@ -10,21 +10,13 @@ class Text extends BladeComponent
 {
     public function __construct(
         public string $value = '',
-        public bool $escape = false,
-        public bool $translate = false,
+        public bool $escaped = false,
+        public bool $translated = false,
     ) {
     }
 
     public function render(): View|Factory
     {
         return view('ui::components.text');
-    }
-
-    /**
-     * Whether the component should be rendered
-     */
-    public function shouldRender(): bool
-    {
-        return ! empty($this->value);
     }
 }