+++ src/Core/Facade.php
@@ -5,16 +5,20 @@ namespace Sikessem\UI\Core;
 use Illuminate\Support\Facades\Facade as BaseFacade;
 
 /**
- * @method static string build(string $name, ComponentSlot $slot)
  * @method static void component(string $class, string $alias = null, bool $anonymous = false)
  * @method static string compress(string $contents)
- * @method static string end()
+ * @method static string close()
  * @method static array|null find(string $name)
  * @method static bool isBlade(string $component, bool $anonymous = false)
+ * @method static bool isBlock(string $tag)
+ * @method static bool isInline(string $tag)
  * @method static bool isLivewire(string $component, bool $anonymous = false)
- * @method static string make(string $name = null, array|\Illuminate\View\ComponentAttributeBag $attributes = [], string|\Illuminate\View\ComponentSlot $slot = null)
- * @method static string renderComponent(string $content, array $data = [], bool $deleteCachedView = true)
- * @method static \Illuminate\Contracts\View\View renderPage(string $content, string $layout = null, \Illuminate\Contracts\Support\Arrayable|array $contentData = [], \Illuminate\Contracts\Support\Arrayable|array $layoutData = [], array $mergeData = [])
+ * @method static bool isOrphan(string $tag)
+ * @method static bool isPaired(string $tag)
+ * @method static string make(string $name, array|\Illuminate\View\ComponentAttributeBag $attributes = [], string|\Illuminate\View\ComponentSlot $slot = null)
+ * @method static string open(string $name, array|ComponentAttributes $attributes = [], string|ComponentSlot $slot = null)
+ * @method static string render(string $template, array $data = [], bool $deleteCachedView = true)
+ * @method static \Illuminate\Contracts\View\View page(string $contentPath, \Illuminate\Contracts\Support\Arrayable|array $contentData = [], string $layoutPath = null, \Illuminate\Contracts\Support\Arrayable|array $layoutData = [], array $mergeData = [])
  *
  * @see \Sikessem\UI\Core\Manager
  */