+++ tests/Feat/UICompilerTest.php
@@ -3,5 +3,5 @@
 namespace Tests\Unit;
 
 it('should compile tag', function () {
-    expect('Hello')->toBeRenderOf('<ui-text value="Hello"/>');
+    expect('Hello')->toBeRenderOf('<s-text value="Hello"/>');
 });