+++ tests/Feat/UIDirectiveTest.php
@@ -1,23 +1,23 @@
 <?php
 
-namespace Tests\Unit;
-
-it('should render HTML element', function () {
-    expect('<div class="app">Hello</div>')->toBeRenderOf('@ui("div", ["class" => "app"], "Hello")');
-    expect('<img src="...">')->toBeRenderOf('@ui("img", ["src" => "..."])');
-});
+namespace Tests\Feat;
 
 it('should render Blade component', function () {
-    expect('Sikessem')->toBeRenderOf('@ui("text", ["value" => "Sikessem"])');
-    expect('<a href="http://localhost">Sikessem</a>')->toBeRenderOf('@ui("link", ["href" => "/", "text" => "Sikessem"])');
+    expect('Sikessem')->toBeRenderOf('@s("text", ["value" => "Sikessem"])');
+    expect('<a href="http://localhost">Sikessem</a>')->toBeRenderOf('@s("link", ["href" => "/", "text" => "Sikessem"])');
 });
 
 it('should render Livewire component', function () {
-    expectRender('@ui("hello-livewire")')->toContain('Hello');
+    expectRender('@s("hello-livewire")')->toContain('Hello');
 });
 
-it('should open HTML tag', function () {
-    expect('<div>Hello World !</div>')
-        ->toBeRenderOf('@ui("div")Hello World !@ui')
-        ->toBeRenderOf('@ui("div")Hello World !@endui');
+it('should render self-closing ui directive', function () {
+    expect('<br/>')->toBeRenderOf('@ui("br")');
+    expect('<img src="..."/>')->toBeRenderOf('@ui("img", ["src" => "..."])');
+});
+
+it('should render open and close ui directive', function () {
+    expect('<div class="app">Hello World !</div>')
+        ->toBeRenderOf('@ui("div", ["class" => "app"], "Hello World !")')
+        ->toBeRenderOf('@ui("div", ["class" => "app"])Hello World !@endui');
 });