+++ CHANGELOG.md
@@ -6,7 +6,7 @@ All notable changes to [Sikessem's UI kit](https://github.com/sikessem/ui) will 
 
 **Full Changelog:** [v0.1.0...v0.2.0](https://github.com/sikessem/ui/compare/v0.1.0...v0.2.0)
 
-- [0.x] Fix end of [blade directive](https://github.com/sikessem/ui#blade-directive)
+- [0.x] Change [blade directive](https://github.com/sikessem/ui#blade-directives)
 - [0.x] Change [**flashes** component](https://github.com/sikessem/ui#flashes-component)
 - [0.x] Change [**flash** component](https://github.com/sikessem/ui#flash-component)
 
@@ -18,13 +18,13 @@ All notable changes to [Sikessem's UI kit](https://github.com/sikessem/ui) will 
 - [0.x] Add [**label** component](https://github.com/sikessem/ui#label-component)
 - [0.x] Add [**menu** component](https://github.com/sikessem/ui#menu-component)
 - [0.x] Add [**button** component](https://github.com/sikessem/ui#button-component)
-- [0.x] Add [blade directive end](https://github.com/sikessem/ui#blade-directive)
+- [0.x] Add [blade directive end](https://github.com/sikessem/ui#blade-directives)
 
 ## [v0.1.0](https://github.com/sikessem/ui/releases/tag/v0.1.0) - 2023-07-17
 
 - [0.x] Add [inline component](https://github.com/sikessem/ui#inline-component)
 - [0.x] Add [user guide](https://github.com/sikessem/ui#-usage)
-- [0.x] Add [blade directive](https://github.com/sikessem/ui#blade-directive)
+- [0.x] Add [blade directives](https://github.com/sikessem/ui#blade-directives)
 - [0.x] Add [**link** component](https://github.com/sikessem/ui#link-component)
 - [0.x] Add [**text** component](https://github.com/sikessem/ui#text-component)
 - [0.x] Add **base-layout** component